backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
google-generativeai==0.3.2
PyPDF2==3.0.1
python-docx==1.1.0
Pillow==10.4.0
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.25.2
sentence-transformers==2.2.2
torch==2.1.2
transformers==4.36.2
reportlab==4.0.7
